Add an offline toggle
